// packages/util/src/declaration-parser/types.ts
export interface ArrowFunctionExpression {
  type: 'ArrowFunctionExpression';
  generator: boolean;
  id: null;
  params: Parameter[];
  // body: BlockStatement | Expression;
  async: boolean;
  // expression: boolean;
}
export interface AssignmentPattern {
  type: 'AssignmentPattern';
  left: BindingName;
  // right: Expression;
}
export type BindingName = BindingPattern | Identifier;
export type BindingPattern = ArrayPattern | ObjectPattern;
interface ClassDeclarationBase {
  id: Identifier | null;
  // body: ClassBody;
  // superClass: LeftHandSideExpression;
}
export interface ClassDeclaration extends ClassDeclarationBase {
  type: 'ClassDeclaration';
}
export interface ClassExpression extends ClassDeclarationBase {
  type: 'ClassExpression';
}
export type DestructuringPattern =
  | ArrayPattern
  | AssignmentPattern
  | Identifier
  // | MemberExpression
  | ObjectPattern
  | RestElement;
export type Expression =
  // Only parse function and class expression for now
  | ArrowFunctionExpression
  | ClassExpression
  | FunctionExpression
  | { type: 'UnknownExpression' };
export interface ExportAllDeclaration {
  type: 'ExportAllDeclaration';
  source: StringLiteral | null;
  exported: Identifier | null;
  // assertions: ImportAttribute[];
}
export type ExportDeclaration =
  | ClassDeclaration
  | ClassExpression
  | FunctionDeclaration
  | VariableDeclaration;
export interface ExportDefaultDeclaration {
  type: 'ExportDefaultDeclaration';
  declaration: ExportDeclaration | Expression;
}
export interface ExportNamedDeclaration {
  type: 'ExportNamedDeclaration';
  declaration: ExportDeclaration | null;
  specifiers: ExportSpecifier[];
  source: StringLiteral | null;
  // assertions: ImportAttribute[];
}
export interface ExportSpecifier {
  type: 'ExportSpecifier';
  local: Identifier;
  exported: Identifier;
}
export interface FunctionDeclaration extends FunctionDeclarationBase {
  type: 'FunctionDeclaration';
  // body: BlockStatement;
}
interface FunctionDeclarationBase {
  id: Identifier | null;
  generator: boolean;
  // expression: boolean;
  async: boolean;
  params: Parameter[];
}
export interface FunctionExpression extends FunctionDeclarationBase {
  type: 'FunctionExpression';
  // body: BlockStatement;
}
// ...
export interface VariableDeclaration {
  type: 'VariableDeclaration';
  declarations: VariableDeclarator[];
  kind: 'const' | 'let' | 'var';
}